Added support for pingService.
This allows for a long running polling of the state of the drone service.
It just streams a DroneReply so that it says whether the drone is
reachable or not. Note that besides a NOK answer there can also be an
exeception. The exception will normally mean the service is not operational
while the NOK means the service is working but cannot reach the drone.