Modifications for Android as the service provider.
Some of the Video names were renamed to LiveVideo to reflect
the distinction between live and recorded video on the drone.

Also some of the Live video structures and calls were removed
as with the concept of video room which the android application
uses we no longer need RTP connection data. This may change in
the future